move memorypool from lock_free to event subfolder
